replace tests to correct test file
